requirements.txt
Flask==3.0.3
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
gunicorn==22.0.0
